Moving SelectExpandDefinition related classes to LogicBuilder.Expressions.Utils.
