+++ LogicBuilder.Expressions.Utils/Expansions/SelectExpandDefinition.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace LogicBuilder.Expressions.Utils.Expansions
+{
+    public class SelectExpandDefinition
+    {
+        public List<string> Selects { get; set; } = new List<string>();
+        public List<SelectExpandItem> ExpandedItems { get; set; } = new List<SelectExpandItem>();
+    }
+}
