+++ LogicBuilder.Expressions.Utils/Expansions/SelectExpandItemFilter.cs
@@ -1,0 +1,10 @@
+﻿using LogicBuilder.Expressions.Utils.ExpressionDescriptors;
+
+namespace LogicBuilder.Expressions.Utils.Expansions
+{
+    public class SelectExpandItemFilter
+    {
+        public IExpressionDescriptor FilterBody { get; set; }
+        public string ParameterName { get; set; }
+    }
+}
