+++ LogicBuilder.Expressions.Utils/Expansions/SelectExpandItemQueryFunction.cs
@@ -1,0 +1,9 @@
+﻿using LogicBuilder.Expressions.Utils.ExpressionDescriptors;
+
+namespace LogicBuilder.Expressions.Utils.Expansions
+{
+    public class SelectExpandItemQueryFunction
+    {
+        public IExpressionDescriptor MethodCallDescriptor { get; set; }
+    }
+}
