+++ LogicBuilder.Structures/Expansions/SelectExpandDefinition.cs
@@ -2,9 +2,9 @@
 
 namespace LogicBuilder.Expressions.Utils.Expansions
 {
-    public class SelectExpandDefinition
-    {
-        public List<string> Selects { get; set; } = new List<string>();
-        public List<SelectExpandItem> ExpandedItems { get; set; } = new List<SelectExpandItem>();
-    }
+    //public class SelectExpandDefinition
+    //{
+    //    public List<string> Selects { get; set; } = new List<string>();
+    //    public List<SelectExpandItem> ExpandedItems { get; set; } = new List<SelectExpandItem>();
+    //}
 }
