+++ LogicBuilder.Structures/Expansions/SelectExpandItemFilter.cs
@@ -2,9 +2,9 @@
 
 namespace LogicBuilder.Expressions.Utils.Expansions
 {
-    public class SelectExpandItemFilter
-    {
-        public IExpressionDescriptor FilterBody { get; set; }
-        public string ParameterName { get; set; }
-    }
+    //public class SelectExpandItemFilter
+    //{
+    //    public IExpressionDescriptor FilterBody { get; set; }
+    //    public string ParameterName { get; set; }
+    //}
 }
